more fixing of the heartbeat bug
I think I got it this time. I identified a way to test the bug: I create an id for the socket connection and pass it as a parameter to the heartbeat function, and when the socket is detected as dead, I make a new id when I create my new socket connection. Then I check which socket id is passed to the heartbeat function on the next go around, and if the id does not match the new one, I remove all event listeners on that connection. This new technique allowed me to experiment by manually modifying the connection id to see what happens. To my surprise, I saw that the new connection was created properly and the old connection seemed to be severed properly, but the heartbeat function never stopped running, and in fact a new instance of it got spun up every time a reconnection happened. After a few dozen reconnections, this resulted in dozens of queries to the relay, some of them simultaneous. Eventually the relay stopped responding to me, which caused the script to bug out and stop running. The new fix prevents makes sure that when a new connection opens, the heartbeat function for the previous connection stops. This should result in a stable connection/reconnection process, thus fixing the bug. Here's hoping.